api-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pinecone-client==3.0.2
openai==1.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.1
